Brings in findpath (ff-findpath binary), greedy MFE search (ff-greedy),
   and the PairTable Display impl from dev_jonjaeg.

   Conflict in pair_table.rs resolved by:
   - Keeping Index<NAIDX>/IndexMut<NAIDX> impls (dev_cotr_pk) over the
     equivalent Index<u16>/IndexMut<u16> impls (dev_jonjaeg)
   - Keeping both extended-bracket tests (dev_cotr_pk) and the new
     test_display_pair_table test (dev_jonjaeg)

+++ crates/ff_structure/src/loop_table.rs
@@ -30,7 +30,7 @@ impl From<&PairTable> for LoopTable {
 
         let mut stack: Vec<(usize, NAIDX)> = Vec::new(); // (closing_idx, loop_id)
         for i in 0..n {
-            match pt[i] {
+            match pt[i as u16] {
                 None => {
                     table[i] = LoopInfo::Unpaired { l: loop_index };
                 }
